creados dtos para CartItem. Implementado el controller para gestionar Cart y los CartItems asociados

+++ ShoppingCart/src/main/kotlin/org/mlc/shoppingcart/dto/product/UpdateProductRequest.kt
@@ -1,10 +1,12 @@
 package org.mlc.shoppingcart.dto.product
 
+import java.math.BigDecimal
+
 data class UpdateProductRequest(
     val name: String?,
     val brand: String?,
     val description: String?,
-    val price: Double?,
+    val price: BigDecimal?,
     val inventory: Int?,
     val categoryName: String?,
 )
